make the DOM rendering less naive
Previously the render method just subscribed directly to the store, and
ran on any state updates, which results in a lot of unnecessary
executions. This change implements a targeted observer method, so that
we can subscribe directly to specific bits of the state. As a result,
rendering is now much more targeted.